+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
+- Remove own userId from heroes in room name
 - Fix server versions not received leading to stuck media downloads
 
 ### Security